backend/requirements.txt
python-dotenv
boto3
fastapi
uvicorn
supabase
httpx
requests
